[agent] Phase 2: hoist ValidatorKind to gaze-types; expose via Recognizer
