+++ crates/gaze-types/Cargo.toml
@@ -21,4 +21,10 @@ path = "src/lib.rs"
 
 [dependencies]
 serde = { workspace = true, features = ["derive"] }
+sha3 = "0.10"
 thiserror = { workspace = true }
+phonenumber = { version = "0.3.9", default-features = false, optional = true }
+
+[features]
+default = []
+phone-parser = ["dep:phonenumber"]
